requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
asyncio==3.4.3
httpx==0.25.0
motor==3.6.0
ntgcalls==2.0.6
pillow==9.5.0
psutil==7.0.0
py-tgcalls==2.2.6
python-dotenv==1.0.1
tgcrypto==1.2.5
youtube-search-python
yt-dlp==2025.8.20
https://github.com/adityahalder/pyrogram/archive/master.zip
